requirements.txt
asgiref==3.6.0
attrs==21.4.0
bcrypt==4.0.1
blessed==1.18.1
boto3==1.26.0
botocore==1.29.0
cached-property==1.5.2
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
colorama==0.4.4
cryptography==39.0.1
Django==3.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-cleanup==5.2.0
django-phonenumber-field==5.1.0
django-session-timeout==0.1.0
django-storages==1.14.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
future==0.18.2
idna==3.4
jmespath==1.0.1
jsonschema==4.17.3
packaging==23.1
paramiko==2.11.0
pathspec==0.11.1
phonenumbers==8.13.18
Pillow==9.5.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.0
semantic-version==2.10.0
six==1.16.0
sqlparse==0.4.3
termcolor==1.1.0
texttable==1.6.4
urllib3==1.26.16
wcwidth==0.2.6
websocket-client==1.5.0
